refactor(test-runners): extract framework detection and AC parsing into src/test-runners/
Introduces src/test-runners/ as the SSOT for all test-framework-aware concerns:
- detector.ts: detectFramework() + Framework type (was inline in verification/parser.ts)
- parser.ts: parseTestOutput() and per-framework sub-parsers (moved from verification/parser.ts)
- ac-parser.ts: parseTestFailures() with framework-triaged AC-ID extraction (extracted from pipeline/stages/acceptance.ts)
- types.ts: TestFailure, TestSummary, TestOutputAnalysis (moved from verification/types.ts)
- index.ts: barrel

Fixes layering: both pipeline/ and verification/ now import from test-runners/ (lower-level
utility) instead of verification/ importing from pipeline/. Adds jest/vitest AC-ID support
as a side effect of the framework triage. verification/parser.ts and verification/types.ts
become thin re-export shims for backward compatibility.
