+++ src/acceptance/hardening.ts
@@ -9,9 +9,9 @@ import type { AgentAdapter } from "../agents/types";
 import { type ModelDef, resolveModelForAgent } from "../config";
 import type { NaxConfig } from "../config";
 import { getSafeLogger } from "../logger";
-import { parseTestFailures } from "../pipeline/stages/acceptance";
 import { savePRD } from "../prd";
 import type { PRD } from "../prd/types";
+import { parseTestFailures } from "../test-runners/ac-parser";
 import { buildAcceptanceRunCommand } from "./generator";
 import { generateFromPRD } from "./generator";
 import { refineAcceptanceCriteria } from "./refinement";
